Safari-611.1.10/LayoutTests/media/modern-media-controls/pip-placard/pip-placard-expected.txt

Testing the PiPPlacard class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Placard with all parameters
PASS placard.children[0].children[0].iconName is Icons.PiPPlacard
PASS placard.children[0].children[1].element.innerText is "This video is playing in picture in picture."
PASS successfullyParsed is true

TEST COMPLETE
